Move keys and address to types package
