+++ src/main/java/com/ltonetwork/client/utils/CryptoUtil.java
@@ -7,12 +7,11 @@ import com.goterl.lazysodium.interfaces.Box;
 import com.goterl.lazysodium.interfaces.GenericHash;
 import com.goterl.lazysodium.interfaces.Sign;
 import com.goterl.lazysodium.utils.LibraryLoader;
-import com.ltonetwork.client.core.Key;
-import com.ltonetwork.client.core.KeyPair;
+import com.ltonetwork.client.types.Key;
+import com.ltonetwork.client.types.KeyPair;
 import com.ltonetwork.client.types.Encoding;
 
 import java.nio.charset.StandardCharsets;
-import java.util.regex.Pattern;
 
 public class CryptoUtil {
